add update groups to the update strategy
this will let people create ordered updates within world
types. for instance, if there are worlds that always operate
based on the data of other worlds, one way to solve this
is to put both types in their own UpdateGroup and make the
dependent worlds ordered before the others.

this also removes a lot of unneeded update code.